requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.22
Faker==8.10.0
python-dotenv==0.19.0
